fix: augmentation of preset default and web sandbox react types
