+++ apps/web-sandbox/src/App.tsx
@@ -43,6 +43,7 @@ function App() {
     codeStyle: {
       display: 'block',
       py: 's',
+      color: 'primary.dark',
     },
   });
